requirements.txt
Django==2.0.6
django-cors-headers==2.1.0
django-filter==1.1.0
djangorestframework==3.7.7
Markdown==2.6.11
mysqlclient==1.3.12
pytz==2018.3
